don't catch RuntimeError in "import clr"
Revert the change added in f765aad4c1f85e994845aedda374999130e07da8
A RuntimeError was raised when uv installed pythonnet in Python 3.14.
However, a RuntimeError is also raised if a .NET runtime is not
installed. pythonnet provides the appropriate error message when
raising the RuntimeError, so let pythonnet raise it.